+++ daemon/internal/image/tarexport/load.go
@@ -16,8 +16,8 @@ import (
 	"github.com/distribution/reference"
 	"github.com/docker/distribution"
 	"github.com/docker/docker/daemon/internal/image"
+	"github.com/docker/docker/daemon/internal/ioutils"
 	"github.com/docker/docker/daemon/internal/layer"
-	"github.com/docker/docker/internal/ioutils"
 	"github.com/docker/docker/pkg/progress"
 	"github.com/docker/docker/pkg/streamformatter"
 	"github.com/docker/docker/pkg/stringid"
