+++ daemon/internal/image/tarexport/save.go
@@ -18,8 +18,8 @@ import (
 	"github.com/docker/distribution"
 	"github.com/docker/docker/daemon/internal/image"
 	v1 "github.com/docker/docker/daemon/internal/image/v1"
+	"github.com/docker/docker/daemon/internal/ioutils"
 	"github.com/docker/docker/daemon/internal/layer"
-	"github.com/docker/docker/internal/ioutils"
 	"github.com/docker/docker/pkg/system"
 	"github.com/moby/go-archive"
 	"github.com/moby/moby/api/types/events"
